python-server/requirements.txt
numpy==2.2.4
scipy==1.15.2
rx==3.2.0
websockets==15.0.1
aws_sdk_bedrock_runtime
mcp==1.8.1
strands-agents
boto3>=1.26.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
requests>=2.28.0
requests-aws4auth>=0.9.0